refactor: migrate user and domain config types to api-clients
