+++ apps/web/src/api-clients/add-ons/favorite/schema/model.ts
@@ -1,4 +1,4 @@
-import type { UserConfigModel } from '@/schema/config/user-config/model';
+import type { UserConfigModel } from '@/api-clients/config/user-config/schema/model';
 
 import type { FavoriteType } from '@/common/modules/favorites/favorite-button/type';
