+++ apps/web/src/api-clients/config/public-config/schema/api-verbs/delete.ts
