+++ apps/web/src/api-clients/config/public-config/schema/type.ts
@@ -1,0 +1,3 @@
+import type { PUBLIC_CONFIG_NAMES } from '@/api-clients/config/public-config/schema/constant';
+
+export type PublicConfigKey = keyof typeof PUBLIC_CONFIG_NAMES;
