+++ apps/web/src/api-clients/config/shared-config/schema/type.ts
@@ -1,4 +1,4 @@
-import type { SHARED_CONFIG_NAMES } from '@/schema/config/shared-config/constant';
+import type { SHARED_CONFIG_NAMES } from '@/api-clients/config/shared-config/schema/constant';
 
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
