+++ apps/web/src/common/modules/favorites/favorite-button/store/favorite-store.ts
@@ -6,10 +6,10 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
-import type { UserConfigDeleteParameters } from '@/schema/config/user-config/api-verbs/delete';
-import type { UserConfigListParameters } from '@/schema/config/user-config/api-verbs/list';
-import type { UserConfigSetParameters } from '@/schema/config/user-config/api-verbs/set';
-import type { UserConfigModel } from '@/schema/config/user-config/model';
+import type { UserConfigDeleteParameters } from '@/api-clients/config/user-config/schema/api-verbs/delete';
+import type { UserConfigListParameters } from '@/api-clients/config/user-config/schema/api-verbs/list';
+import type { UserConfigSetParameters } from '@/api-clients/config/user-config/schema/api-verbs/set';
+import type { UserConfigModel } from '@/api-clients/config/user-config/schema/model';
 
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
 import { useUserStore } from '@/store/user/user-store';
