--- apps/web/src/schema/config/public-config/type.ts
@@ -1,3 +1,0 @@
-import type { PUBLIC_CONFIG_NAMES } from '@/schema/config/public-config/constant';
-
-export type PublicConfigKey = keyof typeof PUBLIC_CONFIG_NAMES;
