+++ apps/web/src/services/advanced/pages/admin/AdminDomainSettingsAutoDormancyConfigurationPage.vue
@@ -11,12 +11,12 @@ import {
 import { getNumberFromString, numberFormatter } from '@cloudforet/utils';
 
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
+import type { DomainConfigListParameters } from '@/api-clients/config/domain-config/schema/api-verbs/list';
+import type { DomainConfigUpdateParameters } from '@/api-clients/config/domain-config/schema/api-verbs/update';
+import { DOMAIN_CONFIG_NAMES } from '@/api-clients/config/domain-config/schema/constant';
+import type { DomainConfigModel } from '@/api-clients/config/domain-config/schema/model';
 import type { CostReportConfigListParameters } from '@/api-clients/cost-analysis/cost-report-config/schema/api-verbs/list';
 import type { CostReportConfigModel } from '@/api-clients/cost-analysis/cost-report-config/schema/model';
-import type { DomainConfigListParameters } from '@/schema/config/domain-config/api-verbs/list';
-import type { DomainConfigUpdateParameters } from '@/schema/config/domain-config/api-verbs/update';
-import { DOMAIN_CONFIG_NAMES } from '@/schema/config/domain-config/constant';
-import type { DomainConfigModel } from '@/schema/config/domain-config/model';
 import { i18n as _i18n } from '@/translations';
