+++ apps/web/src/services/advanced/store/bookmark-page-store.ts
@@ -10,10 +10,10 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
+import type { SharedConfigListParameters } from '@/api-clients/config/shared-config/schema/api-verbs/list';
+import type { SharedConfigModel } from '@/api-clients/config/shared-config/schema/model';
 import type { WorkspaceListParameters } from '@/api-clients/identity/workspace/api-verbs/list';
 import type { WorkspaceModel } from '@/api-clients/identity/workspace/model';
-import type { SharedConfigListParameters } from '@/schema/config/shared-config/api-verbs/list';
-import type { SharedConfigModel } from '@/schema/config/shared-config/model';
 
 import { fetchFavicon } from '@/common/components/bookmark/composables/use-bookmark';
 import type { BookmarkItem } from '@/common/components/bookmark/type/type';
