+++ apps/web/src/services/workspace-home/components/BookmarkBoard.vue
@@ -10,8 +10,8 @@ import {
 import type { MenuItem } from '@cloudforet/mirinae/src/controls/context-menu/type';
 import { CONTEXT_MENU_TYPE } from '@cloudforet/mirinae/src/controls/context-menu/type';
 
+import type { UserConfigModel } from '@/api-clients/config/user-config/schema/model';
 import { ROLE_TYPE } from '@/api-clients/identity/role/constant';
-import type { UserConfigModel } from '@/schema/config/user-config/model';
 import { i18n } from '@/translations';
 
 import { useUserStore } from '@/store/user/user-store';
