+++ apps/web/src/services/workspace-home/pages/WorkspaceHomePage.vue
@@ -5,9 +5,9 @@ import {
 
 import { isEmpty } from 'lodash';
 
+import type { UserConfigModel } from '@/api-clients/config/user-config/schema/model';
 import type { CostDataSourceModel } from '@/api-clients/cost-analysis/data-source/schema/model';
 import { ROLE_TYPE } from '@/api-clients/identity/role/constant';
-import type { UserConfigModel } from '@/schema/config/user-config/model';
 
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
