+++ apps/web/src/store/config/domain-config-store.ts
@@ -5,11 +5,11 @@ import { defineStore } from 'pinia';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
-import type { DomainConfigCreateParameters } from '@/schema/config/domain-config/api-verbs/create';
-import type { DomainConfigListParameters } from '@/schema/config/domain-config/api-verbs/list';
-import type { DomainConfigUpdateParameters } from '@/schema/config/domain-config/api-verbs/update';
-import { DOMAIN_CONFIG_NAMES } from '@/schema/config/domain-config/constant';
-import type { DomainConfigModel } from '@/schema/config/domain-config/model';
+import type { DomainConfigCreateParameters } from '@/api-clients/config/domain-config/schema/api-verbs/create';
+import type { DomainConfigListParameters } from '@/api-clients/config/domain-config/schema/api-verbs/list';
+import type { DomainConfigUpdateParameters } from '@/api-clients/config/domain-config/schema/api-verbs/update';
+import { DOMAIN_CONFIG_NAMES } from '@/api-clients/config/domain-config/schema/constant';
+import type { DomainConfigModel } from '@/api-clients/config/domain-config/schema/model';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
